Rename InputPassword to InputTakePassword
